% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta}
\alias{meta}
\alias{meta<-}
\alias{addMeta}
\title{addMeta
Add or change meta information for a cwl recipe.}
\usage{
meta(cwl)

meta(cwl) <- value

addMeta(
  cwl,
  label = character(),
  doc = character(),
  inputLabels = character(),
  inputDocs = character(),
  outputLabels = character(),
  outputDocs = character(),
  stepLabels = character(),
  stepDocs = character(),
  extensions = list()
)
}
\arguments{
\item{cwl}{\code{cwlProcess} object for data or tool
recipe. \code{cwlWorkflow} object for a pipeline recipe.}

\item{label}{Character string specifying a label for the
recipe. E.g., "bwa align", "gencode annotation".}

\item{doc}{Character string describing the recipe. E.g,
"Align reads to reference genome".}

\item{inputLabels}{Vector of character string, specifying labels
for each input parameter.}

\item{inputDocs}{Vector of character string as descriptions for
each input parameter.}

\item{outputLabels}{Vector of character string, specifying labels
for each output parameter.}

\item{outputDocs}{Vector of character string as descriptions for
each output parameter.}

\item{stepLabels}{Vector of character string, specifying labels for
each step. Use only if \code{cwl} is a \code{cwlWorkflow} object.}

\item{stepDocs}{Vector of character string as description for each
step. Use only if \code{cwl} is a \code{cwlWorkflow} object.}

\item{extensions}{A list of character strings. Can be used to add
meta information about the recipe. Generally, add fields of
information that does not require execution as part of the
recipe evaluation. for information about "author", "url",
"date", "example", use the exact names as list names as shown
in examples, so that they can be correctly passed into
corresponding fields in markdown file when using
\code{meta2md}. Other information can be added as a list element
with arbitrary names.}
}
\value{
\code{meta()}: return a list of all available meta information
for the \code{cwl} object.

\code{addMeta()}: \code{cwlProcess} or \code{cwlWorkflow} object, with added meta
information, which can be returned using \code{meta(cwl)}. Meta
information can be converted into markdown file with \code{meta2md}
function.
}
\description{
addMeta
Add or change meta information for a cwl recipe.
}
\examples{
rcp <- ReUseData::recipeLoad("gencode_annotation")
meta(rcp)
rcp1 <- addMeta(cwl = rcp,
                label = "",
                doc = "An empty description line", 
                inputLabels = c("input label1", "input label2"),
                inputDocs = c("input description 1", "input description 2"), 
                outputLabels = c("output label1"),
                outputDocs = c("output description 1"), 
                extensions = list(author = "recipe author's name",
                                  url = "http://ftp.ebi.ac.uk/pub/databases/gencode/",
                                  date = as.character(Sys.Date()),
                                  example = "An example chunk of code that was coverted into md format"))
meta(rcp1)
cat(meta2md(rcp1))
}
